backend/requirements.txt
# ── Core Framework ─────────────────────────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12

# ── Pydantic & Settings ─────────────────────────────────────────────────────────
pydantic==2.9.2
pydantic-settings==2.4.0

# ── Database ────────────────────────────────────────────────────────────────────
sqlmodel==0.0.22
pymysql==1.1.1
cryptography==43.0.1

# ── Auth & Security ─────────────────────────────────────────────────────────────
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0

# ── HTTP Client (proxy layer) ───────────────────────────────────────────────────
httpx==0.27.2

# ── AI Providers ────────────────────────────────────────────────────────────────
google-generativeai==0.8.3
langchain-core==0.3.10
langchain-google-genai==2.0.1
anthropic==0.34.0
openai==1.51.0

# ── Encryption (AES-256-GCM vault) ─────────────────────────────────────────────
# cryptography already listed above

# ── Email (Resend) ──────────────────────────────────────────────────────────────
resend==2.3.0

# ── Voice Providers (stubbed — install when API keys available) ─────────────────
deepgram-sdk==3.5.0
elevenlabs==1.4.1
websockets==13.1

# ── Environment Variables ───────────────────────────────────────────────────────
python-dotenv==1.0.1

# ── File Handling ───────────────────────────────────────────────────────────────
aiofiles==24.1.0
python-magic==0.4.27

# ── Testing ─────────────────────────────────────────────────────────────────────
pytest==8.3.2
pytest-asyncio==0.24.0
